requirements.txt
beautifulsoup4
boto3
botocore
cryptography
Django
django-cors-headers
django-discover-runner
django-environ
django-role-permissions
django-ses
django-storages
djangorestframework
djangorestframework-simplejwt
djangorestframework-xml
drf-extra-fields
filetype
gunicorn
idna
jmespath
louis
meson
mysqlclient
packaging
ply
pycparser
PyJWT
PyPDF2
python-dateutil
pytz
pyxdg
requests
s3transfer
setproctitle
six
soupsieve
sqlparse
untangle
urllib3
whitenoise
ldap3
django_seed
django-filter
pillow
pandas
beautifulsoup4
openpyxl
django-redis
django-cacheops
django-crontab
bcrypt
django-model-utils
workalendar
channels
channels_redis
uvicorn